Traduzindo, e Retirando parametros não utilizados

+++ main.py
@@ -1,7 +1,7 @@
 from typing import Union
 
 from fastapi import FastAPI
-from schemas import TodoCreate,TodoEdit
+from schemas import TodoCreate, TodoEdit
 from http import HTTPStatus
 
 from fastapi import Depends, FastAPI, HTTPException
@@ -20,14 +20,16 @@ def read_root():
     return {"Hello": "World"}
 
 
-@app.post("/todo/",response_model=TodoCreate)
+@app.post("/todo/")
 def todo_create(
-    todo: TodoCreate, session: Session = Depends(get_session), 
+    todo: TodoCreate,
+    session: Session = Depends(get_session),
 ):
     todo = Todo(title=todo.title, description=todo.description)
     session.add(todo)
     session.commit()
     session.refresh(todo)
+    
     return todo
 
 
@@ -35,44 +37,50 @@ def todo_create(
 def todo_list(session: Session = Depends(get_session)):
     todos = session.scalars(select(Todo)).all()
     return todos
-  
+
+
 @app.get("/todo/{todo_id}")
-def todo(todo_id:int,session: Session = Depends(get_session)):
-  todo = session.scalar(select(Todo).where(Todo.id == todo_id))
-  return todo
+def todo(todo_id: int, session: Session = Depends(get_session)):
+    todo = session.scalar(select(Todo).where(Todo.id == todo_id))
+    
+    if not todo:
+        raise HTTPException(
+            status_code=HTTPStatus.NOT_FOUND, detail='Tarefa Não encontrada.'
+        )
+    
+    return todo
+
 
 @app.patch("/todo/{todo_id}")
-def todo_edit(todo_id:int,todo:TodoEdit,session: Session = Depends(get_session)):
-    
-    db_todo = session.scalar(
-        select(Todo).where(Todo.id == todo_id)
-    )
-    
-    # if not todo:
-    #     raise HTTPException(
-    #         status_code=HTTPStatus.NOT_FOUND, detail='Task not found.'
-    #     )
-    # print(todo.model_dump())
-    # print(**todo.model_dump())
-    
+def todo_edit(todo_id: int, todo: TodoEdit, session: Session = Depends(get_session)):
+
+    db_todo = session.scalar(select(Todo).where(Todo.id == todo_id))
+
+    if not todo:
+        raise HTTPException(
+            status_code=HTTPStatus.NOT_FOUND, detail="Tarefa Não encontrada."
+        )
+
     for key, value in todo.model_dump(exclude_unset=True).items():
         setattr(db_todo, key, value)
 
-    # query = update(Todo).where(Todo.id == todo_id).values(Todo(title=todo.title, description=todo.description))
     session.add(db_todo)
     session.commit()
     session.refresh(db_todo)
+    
     return db_todo
 
 
 @app.delete("/todo/{todo_id}")
-def todo_delete(todo_id:int,session: Session = Depends(get_session)):
-    db_todo = session.scalar(
-        select(Todo).where(Todo.id == todo_id)
-    )
-    
+def todo_delete(todo_id: int, session: Session = Depends(get_session)):
+    db_todo = session.scalar(select(Todo).where(Todo.id == todo_id))
+
+    if not db_todo:
+        raise HTTPException(
+            status_code=HTTPStatus.NOT_FOUND, detail="Tarefa Não encontrada."
+        )
+
     session.delete(db_todo)
     session.commit()
-    
-    return {'message': 'Task has been deleted successfully.'}
 
+    return {"message": "Tarefa Deletada com Sucesso."}
